dele opp i to filer
